Move: get-my-posts user->post
