+++ src/modules/post/dto/index.ts
@@ -3,3 +3,4 @@ export { GetAllPostDto } from "./get-all-post.dto";
 export { IdDto } from "./id.dto";
 export { PostResponseDto } from "./post-response.dto";
 export { SearchPostDto } from "./search-post.dto";
+export { GetMyPostsDto } from "./get-my-posts.dto";
