+++ src/modules/post/interfaces/IPost.service.ts
@@ -3,6 +3,7 @@ import {
   PageResponseDto,
 } from "../../../shared/page";
 import { User } from "../../user/entity/user.entity";
+import { GetMyPostsDto } from "../dto";
 import { GetAllPostDto } from "../dto/get-all-post.dto";
 import { PostResponseDto } from "../dto/post-response.dto";
 import { SearchPostDto } from "../dto/search-post.dto";
@@ -21,6 +22,7 @@ export interface IPostService {
   delete(user: User, id: number): Promise<void>;
   getDetail(user: User, id: number): Promise<PostResponseDto>;
   isValid(id: number): Promise<boolean>;
+  getMyPosts(user: User, pageOptionsDto: GetMyPostsDto): Promise<any>; // TODO: 리턴타입, postService.getUserPosts로
   getAll(
     user: User,
     pageOptionsDto: GetAllPostDto
