+++ src/modules/user/dto/index.ts
@@ -2,7 +2,6 @@ export { CreateUserDto } from "./create-user.dto";
 export { LoginDto } from "./login.dto";
 export { CheckDuplicateNicknameDto } from "./check-duplicate-nickname.dto";
 export { SignUpDto } from "./sign-up.dto";
-export { GetMyPostsDto } from "./get-my-posts.dto";
 export { UserResponseDto } from "./response/user-response.dto";
 export { UserTokenResponseDto } from "./response/user-token-response.dto";
 export { NeedSignUpResponseDto } from "./response/need-signup-response.dto";
